deprecate CreateXXX functions (move to lts)
