+++ src/Meshes/index.ts
@@ -7,6 +7,7 @@ export * from "./trailMesh";
 export * from "./instancedMesh";
 export * from "./linesMesh";
 export * from "./mesh";
+export * from "./mesh.lts";
 export * from "./mesh.vertexData";
 export * from "./meshBuilder";
 export * from "./meshSimplification";
